refactor(analytic): implement Clarity and Google Analytics setup functions in separate modules
